+++ src/main/AST.cpp
@@ -10,6 +10,7 @@ void MemberAccess::accept(ASTVisitor& visitor) { visitor.visit(*this); }
 void GenericSelection::accept(ASTVisitor& visitor) { visitor.visit(*this); }
 void ArrayAccess::accept(ASTVisitor& visitor) { visitor.visit(*this); }
 void ConditionalExpression::accept(ASTVisitor& visitor) { visitor.visit(*this); }
+void CastExpression::accept(ASTVisitor& visitor) { visitor.visit(*this); }
 void AlignofExpression::accept(ASTVisitor& visitor) { visitor.visit(*this); }
 void SizeofExpression::accept(ASTVisitor& visitor) { visitor.visit(*this); }
 void FunctionCall::accept(ASTVisitor& visitor) { visitor.visit(*this); }